Moving claim information into claims.go
